+++ modules/src/ics07_tendermint/header.rs
@@ -1,12 +1,13 @@
 use serde_derive::{Deserialize, Serialize};
 
 use tendermint::block::signed_header::SignedHeader;
-use tendermint::block::Height;
 use tendermint::validator::Set as ValidatorSet;
 
 use crate::ics02_client::client_type::ClientType;
 use crate::ics07_tendermint::consensus_state::ConsensusState;
 use crate::ics23_commitment::commitment::CommitmentRoot;
+use crate::ics24_host::identifier::ChainId;
+use crate::Height;
 
 /// Tendermint consensus header
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
@@ -33,25 +34,24 @@ impl crate::ics02_client::header::Header for Header {
     }
 
     fn height(&self) -> Height {
-        self.signed_header.header.height
+        Height::new(
+            ChainId::chain_version(self.signed_header.header.chain_id.to_string()),
+            u64::from(self.signed_header.header.height),
+        )
     }
-
-    // fn consensus_state(&self) -> &dyn crate::ics02_client::state::ConsensusState {
-    //     &self.consensus_state()
-    // }
 }
 
 #[cfg(test)]
 pub mod test_util {
     use subtle_encoding::hex;
 
-    use crate::ics07_tendermint::header::Header;
-
     use tendermint::block::signed_header::SignedHeader;
-    use tendermint::block::Height;
     use tendermint::validator::Info as ValidatorInfo;
     use tendermint::validator::Set as ValidatorSet;
     use tendermint::{vote, PublicKey};
+
+    use crate::ics07_tendermint::header::Header;
+    use crate::Height;
 
     // TODO: This should be replaced with a ::default() or ::produce().
     // The implementation of this function comprises duplicate code (code borrowed from
@@ -82,7 +82,7 @@ pub mod test_util {
         Header {
             signed_header: shdr,
             validator_set: vs.clone(),
-            trusted_height: Height::from(9_u32),
+            trusted_height: Height::new(0, 9),
             trusted_validator_set: vs,
         }
     }