+++ modules/src/mock_client/header.rs
@@ -1,13 +1,15 @@
-use crate::ics02_client::client_def::AnyHeader;
+use serde_derive::{Deserialize, Serialize};
+use std::convert::{TryFrom, TryInto};
+
+use ibc_proto::ibc::mock::Header as RawMockHeader;
+use tendermint_proto::DomainType;
+
+use crate::ics02_client::client_def::{AnyConsensusState, AnyHeader};
 use crate::ics02_client::client_type::ClientType;
 use crate::ics02_client::error::{self, Error};
 use crate::ics02_client::header::Header;
-
-use ibc_proto::ibc::mock::Header as RawMockHeader;
-use serde_derive::{Deserialize, Serialize};
-use std::convert::{TryFrom, TryInto};
-use tendermint::block::Height;
-use tendermint_proto::DomainType;
+use crate::mock_client::state::MockConsensusState;
+use crate::Height;
 
 #[derive(Copy, Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct MockHeader(pub Height);
@@ -26,8 +28,7 @@ impl TryFrom<RawMockHeader> for MockHeader {
 
         Ok(MockHeader(
             raw.height
-                .unwrap()
-                .epoch_height
+                .ok_or_else(|| error::Kind::InvalidRawHeader)?
                 .try_into()
                 .map_err(|e| error::Kind::InvalidRawHeader.context(e))?,
         ))
@@ -36,12 +37,7 @@ impl TryFrom<RawMockHeader> for MockHeader {
 
 impl From<MockHeader> for RawMockHeader {
     fn from(value: MockHeader) -> Self {
-        RawMockHeader {
-            height: Some(ibc_proto::ibc::client::Height {
-                epoch_number: 0,
-                epoch_height: value.height().value(),
-            }),
-        } // FIXME: This is wrong as it does not take the epoch into account
+        value.into()
     }
 }
 
@@ -64,5 +60,11 @@ impl Header for MockHeader {
 
     fn height(&self) -> Height {
         todo!()
+    }
+}
+
+impl From<MockHeader> for AnyConsensusState {
+    fn from(h: MockHeader) -> Self {
+        AnyConsensusState::Mock(MockConsensusState(h))
     }
 }